requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests
python-dotenv
